+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/CHANGELOG.md
@@ -4,11 +4,15 @@
 
 ### Features Added
 
+- Added an experimental `PlatformManaged` property on `CertificatePolicy` for Azure Key Vault internal usage only. Any calls using this property will fail and it is not recommended to be used at this point.
+
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+- The default service version is now `2026-03-01-preview`.
 
 ## 4.9.0-beta.1 (2026-04-01)